Refactor LoadingPage component location and enhance error handling in useMap hook
