+++ src/pages/Root.tsx
@@ -2,9 +2,9 @@ import { useState } from "react";
 import { Outlet } from "react-router-dom";
 import Modal from "../components/ui/Modal";
 import ModalRootContent from "../components/root/ModalRootContent";
-import LoadingPage from "./LoadingPage";
 import useProperties from "../hooks/useProperties";
 import useMap from "../hooks/useMap";
+import LoadingPage from "../components/ui/LoadingPage";
 
 const Root = () => {
   const [showModal, setShowModal] = useState(true);
@@ -21,7 +21,9 @@ const Root = () => {
 
   return (
     <main>
-      <div ref={mapDiv} className="h-screen w-full fixed top-0 left-0"></div>
+      {mapDiv && properties.length > 0 && (
+        <div ref={mapDiv} className="h-screen w-full fixed top-0 left-0"></div>
+      )}
       {showModal && (
         <Modal>
           <ModalRootContent toggleModal={toggleModal} />
